# formcraft-specs/specs/032-data-export-integration/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: FormCraft Data Export & Integration API
  version: 0.1.0
  description: Admin submission export, recurring email exports, template package portability, integration credentials, and signed webhooks for F32.
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
...
  /admin/integrations/credentials:
    get:
      summary: List integration credentials
      tags: [Integrations]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Integration credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationCredentialListResponse"
    post:
      summary: Create an integration credential and return the secret once
      tags: [Integrations]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegrationCredentialCreate"
      responses:
        "201":
          description: Created credential with one-time secret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationCredentialCreated"
...
  /admin/integrations/credentials/{credential_id}/revoke:
    post:
      summary: Revoke an integration credential immediately
      tags: [Integrations]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CredentialId"
      responses:
        "200":
          description: Revoked credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationCredential"
  /admin/integrations/webhooks:
    get:
      summary: List webhook subscriptions
      tags: [Integrations]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Webhook subscriptions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSubscriptionListResponse"
    post:
      summary: Create a signed webhook subscription
      tags: [Integrations]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookSubscriptionCreate"
      responses:
        "201":
          description: Created webhook subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSubscription"
        "400":
          description: Signing secret required or invalid URL
  /admin/integrations/webhooks/{webhook_id}:
    patch:
      summary: Update a webhook subscription
      tags: [Integrations]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WebhookId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookSubscriptionUpdate"
      responses:
        "200":
          description: Updated webhook subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSubscription"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CredentialId:
      in: path
      name: credential_id
      required: true
      schema:
        type: string
        format: uuid
    WebhookId:
      in: path
      name: webhook_id
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    IntegrationCredentialCreate:
      type: object
      required: [name, scopes]
      properties:
        name:
          type: string
        scopes:
          type: array
          items:
            type: string
        expires_at:
          type: string
          format: date-time
          nullable: true
    IntegrationCredential:
      type: object
      required: [id, name, key_prefix, scopes, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        key_prefix:
          type: string
        scopes:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, revoked, expired]
        expires_at:
          type: string
          format: date-time
          nullable: true
        last_used_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    IntegrationCredentialCreated:
      allOf:
        - $ref: "#/components/schemas/IntegrationCredential"
        - type: object
          required: [secret]
          properties:
            secret:
              type: string
              description: Shown once only
    IntegrationCredentialListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/IntegrationCredential"
    WebhookSubscriptionCreate:
      type: object
      required: [name, event_type, target_url, signing_secret]
      properties:
        name:
          type: string
        event_type:
          type: string
          enum: [form_submitted, form_printed, template_published, batch_completed]
        target_url:
          type: string
          format: uri
        signing_secret:
          type: string
          minLength: 16
    WebhookSubscriptionUpdate:
      type: object
      properties:
        name:
          type: string
        target_url:
          type: string
          format: uri
        signing_secret:
          type: string
          minLength: 16
        status:
          type: string
          enum: [active, paused, disabled]
    WebhookSubscription:
      type: object
      required: [id, name, event_type, target_url, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        event_type:
          type: string
        target_url:
          type: string
        signing_secret_prefix:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, paused, disabled]
        created_at:
          type: string
          format: date-time
    WebhookSubscriptionListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WebhookSubscription"
...